+++ crates/stackable-operator/src/commons/s3/helpers.rs
@@ -11,9 +11,9 @@ use crate::{
 };
 
 use super::{
-    AddS3CredentialVolumesSnafu, AddS3TlsClientDetailsVolumesSnafu, ParseS3EndpointSnafu,
-    RetrieveS3ConnectionSnafu, S3Bucket, S3BucketSpec, S3Connection, S3ConnectionSpec, S3Error,
-    SetS3EndpointSchemeSnafu,
+    AddS3CredentialVolumesSnafu, AddS3TlsClientDetailsVolumesSnafu, AddVolumeMountsSnafu,
+    AddVolumesSnafu, ParseS3EndpointSnafu, RetrieveS3ConnectionSnafu, S3Bucket, S3BucketSpec,
+    S3Connection, S3ConnectionSpec, S3Error, SetS3EndpointSchemeSnafu,
 };
 
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
@@ -79,19 +79,16 @@ impl ResolvedS3Connection {
     ///
     /// * Credentials needed to connect to S3
     /// * Needed TLS volumes
-    ///
-    /// `unique_identifier` needs to be a unique identifier (e.g. in case of trino-operator the name of the catalog),
-    /// so that multiple mounts of the same SecretClass do not produce clashing volumes and volumeMounts.
     pub fn add_volumes_and_mounts(
         &self,
-        unique_identifier: &str,
         pod_builder: &mut PodBuilder,
         container_builders: Vec<&mut ContainerBuilder>,
     ) -> Result<(), S3Error> {
-        let (volumes, mounts) = self.volumes_and_mounts(unique_identifier)?;
-        pod_builder.add_volumes(volumes);
+        let (volumes, mounts) = self.volumes_and_mounts()?;
+        pod_builder.add_volumes(volumes).context(AddVolumesSnafu)?;
         for cb in container_builders {
-            cb.add_volume_mounts(mounts.clone());
+            cb.add_volume_mounts(mounts.clone())
+                .context(AddVolumeMountsSnafu)?;
         }
 
         Ok(())
@@ -99,16 +96,13 @@ impl ResolvedS3Connection {
 
     /// It is recommended to use [`Self::add_volumes_and_mounts`], this function returns you the
     /// volumes and mounts in case you need to add them by yourself.
-    pub fn volumes_and_mounts(
-        &self,
-        unique_identifier: &str,
-    ) -> Result<(Vec<Volume>, Vec<VolumeMount>), S3Error> {
+    pub fn volumes_and_mounts(&self) -> Result<(Vec<Volume>, Vec<VolumeMount>), S3Error> {
         let mut volumes = Vec::new();
         let mut mounts = Vec::new();
 
         if let Some(credentials) = &self.credentials {
             let secret_class = &credentials.secret_class;
-            let volume_name = format!("{unique_identifier}-{secret_class}-s3-credentials");
+            let volume_name = format!("{secret_class}-s3-credentials");
 
             volumes.push(
                 credentials
@@ -116,11 +110,8 @@ impl ResolvedS3Connection {
                     .context(AddS3CredentialVolumesSnafu)?,
             );
             mounts.push(
-                VolumeMountBuilder::new(
-                    volume_name,
-                    format!("{SECRET_BASE_PATH}/{unique_identifier}-{secret_class}"),
-                )
-                .build(),
+                VolumeMountBuilder::new(volume_name, format!("{SECRET_BASE_PATH}/{secret_class}"))
+                    .build(),
             );
         }
 
@@ -137,15 +128,12 @@ impl ResolvedS3Connection {
 
     /// Returns the path of the files containing bind user and password.
     /// This will be None if there are no credentials for this LDAP connection.
-    ///
-    /// `unique_identifier` needs to be a unique identifier (e.g. in case of trino-operator the name of the catalog),
-    /// so that multiple mounts of the same SecretClass do not produce clashing volumes and volumeMounts.
-    pub fn credentials_mount_paths(&self, unique_identifier: &str) -> Option<(String, String)> {
+    pub fn credentials_mount_paths(&self) -> Option<(String, String)> {
         self.credentials.as_ref().map(|bind_credentials| {
             let secret_class = &bind_credentials.secret_class;
             (
-                format!("{SECRET_BASE_PATH}/{unique_identifier}-{secret_class}/accessKey"),
-                format!("{SECRET_BASE_PATH}/{unique_identifier}-{secret_class}/secretKey"),
+                format!("{SECRET_BASE_PATH}/{secret_class}/accessKey"),
+                format!("{SECRET_BASE_PATH}/{secret_class}/secretKey"),
             )
         })
     }
@@ -214,7 +202,7 @@ mod tests {
             credentials: None,
             tls: TlsClientDetails { tls: None },
         };
-        let (volumes, mounts) = s3.volumes_and_mounts("lakehouse").unwrap();
+        let (volumes, mounts) = s3.volumes_and_mounts().unwrap();
 
         assert_eq!(s3.endpoint().unwrap(), Url::parse("http://minio").unwrap());
         assert_eq!(volumes, vec![]);
@@ -239,7 +227,7 @@ mod tests {
                 }),
             },
         };
-        let (mut volumes, mut mounts) = s3.volumes_and_mounts("lakehouse").unwrap();
+        let (mut volumes, mut mounts) = s3.volumes_and_mounts().unwrap();
 
         assert_eq!(
             s3.endpoint().unwrap(),
@@ -250,10 +238,7 @@ mod tests {
         assert_eq!(mounts.len(), 1);
         let mount = mounts.remove(0);
 
-        assert_eq!(
-            &volume.name,
-            "lakehouse-ionos-s3-credentials-s3-credentials"
-        );
+        assert_eq!(&volume.name, "ionos-s3-credentials-s3-credentials");
         assert_eq!(
             &volume
                 .ephemeral
@@ -271,15 +256,12 @@ mod tests {
         );
 
         assert_eq!(mount.name, volume.name);
+        assert_eq!(mount.mount_path, "/stackable/secrets/ionos-s3-credentials");
         assert_eq!(
-            mount.mount_path,
-            "/stackable/secrets/lakehouse-ionos-s3-credentials"
-        );
-        assert_eq!(
-            s3.credentials_mount_paths("lakehouse"),
+            s3.credentials_mount_paths(),
             Some((
-                "/stackable/secrets/lakehouse-ionos-s3-credentials/accessKey".to_string(),
-                "/stackable/secrets/lakehouse-ionos-s3-credentials/secretKey".to_string()
+                "/stackable/secrets/ionos-s3-credentials/accessKey".to_string(),
+                "/stackable/secrets/ionos-s3-credentials/secretKey".to_string()
             ))
         );
     }
@@ -297,7 +279,7 @@ mod tests {
                 }),
             },
         };
-        let (volumes, mounts) = s3.volumes_and_mounts("lakehouse").unwrap();
+        let (volumes, mounts) = s3.volumes_and_mounts().unwrap();
 
         assert_eq!(
             s3.endpoint().unwrap(),